+++ src/00_raster.cpp
@@ -57,8 +57,7 @@ List warp_in_memory_gdal_cpp(CharacterVector dsn,
                              CharacterVector resample,
                              LogicalVector silent,
                              CharacterVector band_output_type, 
-                             CharacterVector warp_options, 
-                             CharacterVector transformation_options, 
+                     
                              CharacterVector options) {
   return gdalwarpmem::gdal_warp_in_memory(dsn,
                                           source_WKT,
@@ -70,6 +69,5 @@ List warp_in_memory_gdal_cpp(CharacterVector dsn,
                                           resample,
                                           silent,
                                           band_output_type, 
-                                          warp_options, 
-                                          transformation_options, options);
+                                        options);
 }